move database files to their new home
