+++ database/models/databaseJustCreatedSymbol.js
