+++ database/models/serializeArrayForDB.js
