+++ database/models/serializeObjectForDB.js
