--- postgresDB/Dockerfile
@@ -1,5 +1,0 @@
-FROM postgres:9.4
-
-MAINTAINER David Bruant <bruant.d@gmail.com>
-
-VOLUME  ["/etc/postgresql", "/var/log/postgresql", "/var/lib/postgresql", "/var/lib/postgresql/data"]
