+++ test-db.js
@@ -2,7 +2,7 @@
 
 require('./ES-mess/index.js');
 
-var Expressions = require('./postgresDB/index');
+var Expressions = require('./database/management/index');
 
 console.log('node env', process.env.NODE_ENV);
